tests, expecter: Centralize expecter helpers under expecter.go and login.go
There are some generic useful expecter code at ping.go
and utils.go this move that into its own expecter.go file
and also creates a function `VmiConsoleRunCommand` to do
a simple run and check expecter execution at na already logged in VMI.

It also move all the `LoggedIn` functions into login.go
